Working mock assert
This commit contains a working mock assert. This way, we can test whether an assert has been called using a mock object.

However, problems arise when we are testing an unhappy scenario. For example, we want to test that a pointer argument to a function is not NULL. We will assert that argument. In the mock implementation of the assert, we will do mock_c()->actualCall and then return.

This means that the production code will keep running past the assert. So when we try to dereference that NULL pointer later in the function, the whole unit test executable crashes.

Miro Samek brought this up and implemented a plugin (which I could not find). However, looks like it was never integrated into CppUTest. In the future, I can consider expanding CppUTest to include this functionality.

The desired behavior is that when an assert fails, the production code is stopped immediately and does not run further. Also, if an assert is called when it should not have been (we are testing a happy scenario), the test should fail too.

There is an old github issue about this and an old post in google groups. I have both links in Notion.