Modularize CMake and split test executable
For temperature_value tests, we need to define a mock value_holder. This means that the real value_holder cannot be linked into the executable.

Two test executables are defined: one for testing value_holder using the real value_holder implementation, and another for testing temperature_value using a fake value_holder.

CMake build is refactored to be more modular.