Variable requirement allocator WIP
WIP implementation of variable_requirement_allocator that uses the Zephyr memory blocks allocator. I was planning to build and test this for the off-target unit testing, but realized that it does not make sense to test zephyr-dependent code off-target.

This commit can be used as a starting point/example when I start implementing the variable requirement allocator for the target hardware.