current_temperature module and tests
The purpose of this module is to abstract away the single instance of temperature_value that we will use everywhere in the code.

This means that temperature_variable_requirement will not have to worry about where to get the temperature_value instance to call get() on, it will simply call current_temperature_get() and get the temperature it needs for the variable evaluation.