variable_requirement_create() initializes all members of VariableRequirementStruct
The subclasses pass the required member values when calling variable_requirement_create(). This is safer than the subclasses initializing the members directly in their create() functions - this creates more room for error. The function signature also enforces that all required parameters are passed, so all required members will be initialized.