Refactor tests to use fake variable requirement allocator
Before, we had a mock variable requirement allocator that was not actually a mock, it just returned the same piece of memory always. It did not register expected calls.

Now, we switch to having an actual mocked variable requirement allocator. The test itself decides what the variable requirement allocator returns to CUT. The test uses fake_variable_requirement_allocator to allocate the memory and pass it to CUT. fake_variable_requirement_allocator uses our memory block allocator to accomplish that.